refactor : 전략 패턴에서 응집도를 높이기 위해 메서드 매개변수 타입 변경 (#37)
