+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/favorite/repository/customerbakeryfavorite/strategy/BakeryFavoriteSortStrategy.java
@@ -1,14 +1,12 @@
 package com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.strategy;
 
-import com.breaditnow.common.util.GeoPoint;
-import com.breaditnow.domain.domain.bakery.entity.QBakery;
-import com.breaditnow.domain.domain.favorite.entity.QCustomerBakeryFavorite;
 import com.querydsl.core.types.OrderSpecifier;
+import com.querydsl.core.types.dsl.NumberExpression;
 
 public interface BakeryFavoriteSortStrategy {
-	OrderSpecifier<?> getOrderSpecifier(QBakery bakery, QCustomerBakeryFavorite customerBakeryFavorite);
+	OrderSpecifier<?> getOrderSpecifier();
 
 	// 기본 초기화 메서드
-	default void initialize(GeoPoint geoPoint) {
+	default void initialize(NumberExpression<Double> expression) {
 	}
 }
