+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/favorite/repository/customerbakeryfavorite/strategy/LatestBakeryFavoriteSortStrategy.java
@@ -1,15 +1,15 @@
 package com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.strategy;
 
+import static com.breaditnow.domain.domain.favorite.entity.QCustomerBakeryFavorite.*;
+
 import org.springframework.stereotype.Component;
 
-import com.breaditnow.domain.domain.bakery.entity.QBakery;
-import com.breaditnow.domain.domain.favorite.entity.QCustomerBakeryFavorite;
 import com.querydsl.core.types.OrderSpecifier;
 
 @Component
 public class LatestBakeryFavoriteSortStrategy implements BakeryFavoriteSortStrategy {
 	@Override
-	public OrderSpecifier<?> getOrderSpecifier(QBakery bakery, QCustomerBakeryFavorite customerBakeryFavorite) {
+	public OrderSpecifier<?> getOrderSpecifier() {
 		return customerBakeryFavorite.modifiedAt.desc();
 	}
 }
