+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/favorite/repository/customerbakeryfavorite/strategy/PopularBakeryFavoriteSortStrategy.java
@@ -1,9 +1,10 @@
 package com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.strategy;
 
+import static com.breaditnow.domain.domain.bakery.entity.QBakery.*;
+import static com.breaditnow.domain.domain.favorite.entity.QCustomerBakeryFavorite.*;
+
 import org.springframework.stereotype.Component;
 
-import com.breaditnow.domain.domain.bakery.entity.QBakery;
-import com.breaditnow.domain.domain.favorite.entity.QCustomerBakeryFavorite;
 import com.querydsl.core.types.Expression;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.Expressions;
@@ -13,7 +14,7 @@ import com.querydsl.jpa.JPAExpressions;
 @Component
 public class PopularBakeryFavoriteSortStrategy implements BakeryFavoriteSortStrategy {
 	@Override
-	public OrderSpecifier<?> getOrderSpecifier(QBakery bakery, QCustomerBakeryFavorite customerBakeryFavorite) {
+	public OrderSpecifier<?> getOrderSpecifier() {
 		Expression<Long> favoriteCountExpr = JPAExpressions
 			.select(customerBakeryFavorite.count())
 			.from(customerBakeryFavorite)
